refactor: split visualization utilities
